滤波算法整合进程序。
（1）滤波算法中，运动模型的R和测量模型Q需要进一步确定。
（2）矩阵库阵操蛋！

+++ model/TargetGroup.h
@@ -38,7 +38,7 @@ public:
     SingleTarget* getTargetByID(int id);
 
     // 获取当前集群中的目标数量
-    int getTargetCount();
+    size_t getTargetCount();
 
     // 构造一个副本
     TargetGroup* clone();
